app/pyproject.toml
[project]
name = "supasuge-com"
version = "0.1.0"
description = "Personal portfolio and blog"
requires-python = ">=3.12"

dependencies = [
  # Core Flask
  "Flask==3.1.2",
  "flask-sqlalchemy==3.1.1",
  "SQLAlchemy==2.0.45",
  "python-dotenv==1.2.1",
  # Content processing
  "python-frontmatter==1.1.0",
  "markdown",
  "bleach==6.3.0",
  "Pygments==2.19.2",
  # CSRF protection
  "flask-wtf>=1.2.2",
  # Rate limiting
  "Flask-Limiter>=3.8.0",
  "redis>=5.2.0",
  # Analytics
  "geoip2==4.8.1",
  # Production server
  "gunicorn==23.0.0",
  "paramiko>=4.0.0",
  "pycryptodome>=3.23.0",
  "pyperclip>=1.11.0",
  "pandas>=3.0.2",
  "jinja2>=3.1.6",
  "colorama>=0.4.6",
  "openai>=2.32.0",
]

[dependency-groups]
dev = [
  "pylint==4.0.4",
  "mypy==1.13.0",
  "types-bleach",
  "types-Markdown",
  "pytest>=9.0",
]

[tool.uv]
managed = true

[tool.setuptools]
py-modules = ["app", "models", "config", "content_indexer", "content_sync", "security", "wsgi", "database", "extensions", "startup", "content_archive"]

[tool.setuptools.packages.find]
where = ["."]
include = ["blueprints*", "auth*", "services*", "tasks*"]
exclude = ["tests*", "migrations*", "static*", "templates*", "content*", "keys*", "nginx*", "public*", "instance*", "scripts*"]
